Server,Client
Added validation and fixed list-departure.html

+++ WebApp/WebApp/FrontEndApp/FrontEndApp/src/app/components/view/LineAdministration/line-list/line-list.component.ts
@@ -29,7 +29,9 @@ export class LineListComponent implements OnInit {
     }
   }
 onRemove(){
+  if(this.removeLineForm.valid){
   this.lineService.removeLine(this.removeLineForm.value.LineId)
   this.lineService.refreshLines()
+  }
 }
 }
